added work from evening of 4_16_19

+++ Bryan_CCS_Workspace/MSP430_Master_SPI/PEC.c
@@ -27,7 +27,7 @@ void init_PEC15_Table()
     }
 }
 
-uint16_t pec15 (int *data , int len)
+uint16_t pec15 (uint8_t *data , int len)
 {
     uint16_t remainder, address;
     int i;
